# public/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
...
			"ProductImageDto": {
				"description": "DTO pro zobrazení obrázku produktu\nObsahuje všechny properties obrázku produktu včetně ID a timestamps",
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"productId": {
						"type": "number",
						"format": "double"
					},
					"url": {
						"type": "string"
					},
					"title": {
						"type": "string"
					},
					"altText": {
						"type": "string"
					},
					"sortOrder": {
						"type": "number",
						"format": "double"
					},
					"width": {
						"type": "number",
						"format": "double"
					},
					"height": {
						"type": "number",
						"format": "double"
					},
					"imageType": {
						"type": "string"
					},
					"isMain": {
						"type": "boolean"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"id",
					"productId",
					"url",
					"sortOrder",
					"isMain",
					"createdAt",
					"updatedAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CreateProductImageDto": {
				"description": "Data Transfer Object (DTO) pro vytvoření nového obrázku produktu",
				"properties": {
					"productId": {
						"type": "number",
						"format": "double"
					},
					"url": {
						"type": "string"
					},
					"title": {
						"type": "string"
					},
					"altText": {
						"type": "string"
					},
					"sortOrder": {
						"type": "number",
						"format": "double"
					},
					"width": {
						"type": "number",
						"format": "double"
					},
					"height": {
						"type": "number",
						"format": "double"
					},
					"imageType": {
						"type": "string"
					},
					"isMain": {
						"type": "boolean"
					}
				},
				"required": [
					"productId",
					"url"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UpdateProductImageDto": {
				"description": "DTO pro aktualizaci existujícího obrázku produktu\nVšechny properties jsou volitelné",
				"properties": {
					"productId": {
						"type": "number",
						"format": "double"
					},
					"url": {
						"type": "string"
					},
					"title": {
						"type": "string"
					},
					"altText": {
						"type": "string"
					},
					"sortOrder": {
						"type": "number",
						"format": "double"
					},
					"width": {
						"type": "number",
						"format": "double"
					},
					"height": {
						"type": "number",
						"format": "double"
					},
					"imageType": {
						"type": "string"
					},
					"isMain": {
						"type": "boolean"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"DbErrorResponse": {
				"description": "Rozhraní pro odpověď s informací o chybě databáze",
				"properties": {
					"success": {
						"type": "boolean",
						"enum": [
							false
						],
						"nullable": false
					},
					"message": {
						"type": "string"
					},
					"error": {
						"type": "string"
					},
					"recovered": {
						"type": "boolean"
					}
				},
				"required": [
					"success",
					"message",
					"error",
					"recovered"
				],
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {
			"jwt": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT"
			}
		}
	},
	"info": {
		"title": "express-oracle-api",
		"version": "1.0.0",
		"description": "REST API s Express.js, TypeORM, Zod a Oracle",
		"contact": {}
	},
	"paths": {
...
		"/product-images": {
			"get": {
				"operationId": "GetProductImages",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"totalPages": {
											"type": "number",
											"format": "double"
										},
										"limit": {
											"type": "number",
											"format": "double"
										},
										"page": {
											"type": "number",
											"format": "double"
										},
										"total": {
											"type": "number",
											"format": "double"
										},
										"data": {
											"items": {
												"$ref": "#/components/schemas/ProductImageDto"
											},
											"type": "array"
										}
									},
									"required": [
										"totalPages",
										"limit",
										"page",
										"total",
										"data"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"description": "Získá seznam všech obrázků produktů s podporou filtrování, řazení a stránkování",
				"tags": [
					"Obrázky produktů"
				],
				"security": [],
				"parameters": []
			},
			"post": {
				"operationId": "CreateProductImage",
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ProductImageDto"
								}
							}
						}
					}
				},
				"description": "Vytvoří nový obrázek produktu",
				"tags": [
					"Obrázky produktů"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"description": "Data pro vytvoření obrázku",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateProductImageDto",
								"description": "Data pro vytvoření obrázku"
							}
						}
					}
				}
			}
		},
		"/product-images/{id}": {
			"get": {
				"operationId": "GetProductImage",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ProductImageDto"
								}
							}
						}
					}
				},
				"description": "Získá obrázek produktu podle ID",
				"tags": [
					"Obrázky produktů"
				],
				"security": [],
				"parameters": [
					{
						"description": "ID obrázku",
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"put": {
				"operationId": "UpdateProductImage",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ProductImageDto"
								}
							}
						}
					}
				},
				"description": "Aktualizuje existující obrázek produktu",
				"tags": [
					"Obrázky produktů"
				],
				"security": [],
				"parameters": [
					{
						"description": "ID obrázku",
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"description": "Data pro aktualizaci obrázku",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateProductImageDto",
								"description": "Data pro aktualizaci obrázku"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteProductImage",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"description": "Odstraní obrázek produktu",
				"tags": [
					"Obrázky produktů"
				],
				"security": [],
				"parameters": [
					{
						"description": "ID obrázku",
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/product-images/product/{productId}": {
			"get": {
				"operationId": "GetProductImagesByProductId",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/ProductImageDto"
									},
									"type": "array"
								}
							}
						}
					},
					"500": {
						"description": "Chyba při načítání obrázků produktu",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/DbErrorResponse"
								}
							}
						}
					}
				},
				"description": "Získá všechny obrázky pro konkrétní produkt",
				"tags": [
					"Obrázky produktů"
				],
				"security": [],
				"parameters": [
					{
						"description": "ID produktu",
						"in": "path",
						"name": "productId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/product-images/product/{productId}/main": {
			"get": {
				"operationId": "GetMainProductImage",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/ProductImageDto"
										}
									],
									"nullable": true
								}
							}
						}
					},
					"500": {
						"description": "Chyba při načítání hlavního obrázku produktu",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/DbErrorResponse"
								}
							}
						}
					}
				},
				"description": "Získá hlavní obrázek produktu",
				"tags": [
					"Obrázky produktů"
				],
				"security": [],
				"parameters": [
					{
						"description": "ID produktu",
						"in": "path",
						"name": "productId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/product-images/{id}/main/product/{productId}": {
			"put": {
				"operationId": "SetMainImage",
				"responses": {
					"200": {
						"description": "Obrázek byl nastaven jako hlavní",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										},
										"success": {
											"type": "boolean"
										}
									},
									"required": [
										"message",
										"success"
									],
									"type": "object"
								}
							}
						}
					},
					"404": {
						"description": "Obrázek nebyl nalezen",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/DbErrorResponse"
								}
							}
						}
					},
					"500": {
						"description": "Chyba při nastavování hlavního obrázku",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/DbErrorResponse"
								}
							}
						}
					}
				},
				"description": "Nastaví obrázek jako hlavní pro konkrétní produkt",
				"tags": [
					"Obrázky produktů"
				],
				"security": [],
				"parameters": [
					{
						"description": "ID obrázku",
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"description": "ID produktu",
						"in": "path",
						"name": "productId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/product-images/sort-order": {
			"put": {
				"operationId": "UpdateSortOrder",
				"responses": {
					"200": {
						"description": "Pořadí obrázků úspěšně aktualizováno",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										},
										"success": {
											"type": "boolean"
										}
									},
									"required": [
										"message",
										"success"
									],
									"type": "object"
								}
							}
						}
					},
					"500": {
						"description": "Chyba při aktualizaci pořadí obrázků",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/DbErrorResponse"
								}
							}
						}
					}
				},
				"description": "Aktualizuje pořadí zobrazení obrázků",
				"tags": [
					"Obrázky produktů"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"imageIds": {
										"items": {
											"type": "number",
											"format": "double"
										},
										"type": "array"
									}
								},
								"required": [
									"imageIds"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/product-images/seed": {
			"post": {
				"operationId": "SeedTestData",
				"responses": {
					"201": {
						"description": "Testovací data úspěšně seedována",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										},
										"success": {
											"type": "boolean"
										}
									},
									"required": [
										"message",
										"success"
									],
									"type": "object"
								}
							}
						}
					},
					"500": {
						"description": "Chyba při seedování testovacích dat",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/DbErrorResponse"
								}
							}
						}
					}
				},
				"description": "Seedování testovacích dat obrázků produktů",
				"tags": [
					"Obrázky produktů"
				],
				"security": [],
				"parameters": []
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}